evaluation/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.31.0",
    "hpsv2>=1.2.0",
    "image-reward>=1.5",
    "numpy>=2.2.6",
    "open-clip-torch>=2.32.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "protobuf>=3.20.3",
    "sentencepiece>=0.2.1",
    "torch>=2.1.0",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

